@@ -15,44 +15,43 @@ dependencies:
     sdk: flutter
   intl: ^0.17.0
 
-  flutter_bloc: ^7.0.0
-  bloc: ^7.0.0
-  rxdart: ^0.26.0
+  flutter_bloc: ^8.0.1
+  bloc: ^8.0.2
 
-  dio: ^4.0.0
-  sembast: ^3.0.2
-  sembast_web: ^2.0.0
+  dio: ^4.0.4
+  sembast: ^3.1.1+1
+  sembast_web: ^2.0.1+1
 
-  uuid: ^3.0.4
-  flutter_svg: ^0.22.0
+  uuid: ^3.0.5
+  flutter_svg: ^1.0.0
 
-  google_fonts: ^2.0.0
-  font_awesome_flutter: ^9.0.0
-  cupertino_icons: ^1.0.3
+  google_fonts: ^2.2.0
+  font_awesome_flutter: ^9.2.0
+  cupertino_icons: ^1.0.4
 
   qr_flutter: ^4.0.0
-  qr_code_scanner: ^0.4.0
+  qr_code_scanner: ^0.6.1
 
   path: ^1.8.0
-  path_provider: ^2.0.1
-  flutter_secure_storage: ^4.2.0
+  path_provider: ^2.0.8
+  flutter_secure_storage: ^5.0.2
 
   js: ^0.6.3
 
-  ed25519_hd_key: ^2.0.0
+  ed25519_hd_key: ^2.2.0
   bip39: ^1.0.6
 
   didkit:
     path: ../didkit/lib/flutter
 
 
-  flutter_markdown: ^0.6.2
+  flutter_markdown: ^0.6.9
   url_launcher: ^6.0.17
-  logging: ^1.0.1
+  logging: ^1.0.2
   shared_preferences: ^2.0.6
 
   json_annotation: ^4.0.1
-  package_info_plus: ^1.0.6
+  package_info_plus: ^1.3.0
   bip32: ^2.0.0
   hex: ^0.2.0
   secp256k1: ^0.3.0
@@ -66,9 +65,8 @@ dev_dependencies:
     sdk: flutter
   integration_test:
     sdk: flutter
-  #intl_translation: ^0.17.10
   pedantic: ^1.11.0
-  build_runner: ^2.0.5
+  build_runner: ^2.1.7
   json_serializable: ^6.1.3
   flutter_launcher_icons: "^0.9.2"
 flutter_icons:
